Move NonZero shape inference into the reduce submodule
Move the NonZero operator, plus its test, from ops.rs into ops/reduce.rs,
matching how this operator is grouped in src/ops.rs.
